requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
google-generativeai==0.3.2